+++ CHANGELOG.md
@@ -6,6 +6,7 @@ _WIP_
 **Cosmetic:**
 - `circumflex` now turn off all submission text highlighting by default
 - Add `ERROR` and `SUCCESS` labels to some messages
+- Change `YC S/WXX` labels
 
 **Bugfixes:**
 - Fixed a bug where JSON errors were not handled properly